pyproject.toml
[project]
name = "aidotnavigation"
version = "0.1.0"
description = "A simple repository showcasing AI dot-notation traversal through (potentially) large objects via Llama Index and Function Calling"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "asyncio>=3.4.3",
    "llama-index>=0.12.37",
    "llama-index-llms-openai>=0.3.44",
    "llama-index-utils-workflow>=0.3.2",
    "loguru>=0.7.3",
    "nest-asyncio>=1.6.0",
    "openai>=1.82.0",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "typer>=0.16.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.metadata]
allow-direct-references = true


[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E", # pycodestyle
    "F", # pyflakes
    "I", # isort
]

[project.scripts]
test = "scripts.test:app"
